Rename to maplibre-rs except for apple stuff
