+++ .gitignore
@@ -18,7 +18,7 @@ dist/
 
 # Cache by reqwest-middleware-cache
 *reqwest*cache*
-*mapr*cache*
+*maplibre*cache*
 
 logs/
