+++ README.md
@@ -1,9 +1,9 @@
-<h1 align="center">mapr</h1>
+<h1 align="center">maplibre-rs</h1>
 <p align="center">
   <img width="200px" alt="Logo" src="https://">
 </p>
 <div align="center">
-  <img width="300px" src="docs/src/figures/mapr-ios.png" alt="preview">
+  <img width="300px" src="docs/src/figures/maplibre-rs-ios.png" alt="preview">
 </div>
 <div align="center">
   <strong>Native Maps for Web, Mobile and Linux</strong>
@@ -15,12 +15,12 @@
 <div align="center">
   <img src="https://img.shields.io/badge/stability-experimental-orange.svg?style=flat-square" 
       alt="Stability" />
-  <a href="https://github.com/maxammann/mapr/actions/workflows/rust.yml">    
-    <img src="https://github.com/maxammann/mapr/actions/workflows/rust.yml/badge.svg"
+  <a href="https://github.com/maplibre/maplibre-rs/actions/workflows/rust.yml">    
+    <img src="https://github.com/maplibre/maplibre-rs/actions/workflows/rust.yml/badge.svg"
         alt="Build status" /> 
   </a>
-  <a href="https://matrix.to/#/#mapr:matrix.org">    
-    <img src="https://img.shields.io/static/v1?label=Space&message=%23mapr&color=blue&logo=matrix"
+  <a href="https://matrix.to/#/#maplibre:matrix.org">    
+    <img src="https://img.shields.io/static/v1?label=Space&message=%23maplibre&color=blue&logo=matrix"
         alt="Build status" />
   </a>
 </div>
@@ -31,13 +31,13 @@
       Example
     </a>
     <span> | </span>
-    <a href="https://maxammann.org/mapr/docs">
+    <a href="https://maxammann.org/maplibre-rs/docs">
       Book
     </a> | </span>
-    <a href="https://maxammann.org/mapr/api-docs/mapr/">
+    <a href="https://maxammann.org/maplibre-rs/api-docs/maplibre/">
       API
     </a> | </span>
-    <a href="https://matrix.to/#/#mapr:matrix.org">
+    <a href="https://matrix.to/#/#maplibre:matrix.org">
       Chat in Matrix Space
     </a>
   </h3>
@@ -53,16 +53,16 @@ next-gen mapping solution.
 
 ## Description
 
-mapr is a portable and performant vector maps renderer. We aim to support the web, mobile and desktop applications. This
+maplibre-rs is a portable and performant vector maps renderer. We aim to support the web, mobile and desktop applications. This
 is achieved by the novel [WebGPU](https://www.w3.org/TR/webgpu/) specification. Plenty of native implementations are
 already implementing this specification. On the web it is implemented by Firefox, Chrome and Safari. There are also
-standalone implementations which directly use Vulkan, OpenGL or Metal as a backend. Those backends allow mapr to run on
+standalone implementations which directly use Vulkan, OpenGL or Metal as a backend. Those backends allow maplibre-rs to run on
 mobile and desktop applications.
 
 Rust is used as a Lingua-franka on all platforms. This is made possible by WebAssembly which allows us to use Rust for
 web development.
 
-The goal of mapr is to render maps in order to visualize data. Right now the goal of mapr is not to replace existing
+The goal of maplibre-rs is to render maps in order to visualize data. Right now the goal of maplibre-rs is not to replace existing
 vector map renderers like Google Maps, Apple Maps or MapLibre. The current implementation serves as a proof-of-concept
 of the used technology stack. It is unclear whether the high-performance requirements of rendering maps using vector
 graphics are achievable using the current stack.
@@ -86,9 +86,12 @@ https://user-images.githubusercontent.com/905221/163552617-5db04c66-23e3-4915-87
 * Rendering Text
 * Per-Feature Rendering
 * Rendering:
+    * Labels
+    * Symbols
     * Raster data
     * 3D terrain
     * Hill-shade (DEM)
+* Collision detection
 * Support for:
     * GeoJSON
 * API for:
@@ -96,31 +99,12 @@ https://user-images.githubusercontent.com/905221/163552617-5db04c66-23e3-4915-87
     * Swift
     * Java/Kotlin
 
-## Repository Layout
-
-(Note: to be changed)
-
-```bash
-.
-├── docs                # Documentation for mapr
-├── src                 # The source code of the mapr library
-├── libs                # Libraries which will eventually be published as separate crates
-│   ├── mbtiles         # Library for extracting .mbtiles files
-│   ├── style_spec      # Library for interpreting MapLibre style specifications
-│   └── wgsl_validate   # Library for validating WGSL shaders
-├── apple               # Platform specific files for Apple (iOS and MacOS)
-├── web                 # Platform specific files for Web (WebGL and WebGPU)
-├── benches             # Benchmarks for specific parts of the library
-├── examples            # Examples which can be run
-└── test-data           # Geo data which can be used for tests (Usually as .mbtiles)
-```
-
 ## Building & Running
 
 Now, to clone the project:
 
 ```bash
-git clone --recursive git@github.com/maxammann/mapr
+git clone --recursive git@github.com/maplibre/maplibre-rs
 ```
 
 and then build it for running on a desktop:
@@ -136,7 +120,7 @@ cargo run --example desktop --
 ```
 
 More information about building for different platforms can be
-found [here](https://maxammann.org/mapr-docs/building.html).
+found [here](https://maxammann.org/maplibre-rs/docs/development-guide/building.html).
 
 > __Note for Mac__: Before opening the XCode project, you need to build manually using the following command:
 > `cargo build --target aarch64-apple-darwin --lib`
