+++ benches/tessellation.rs
@@ -1,7 +1,7 @@
 use criterion::{criterion_group, criterion_main, Criterion};
 use lyon::tessellation::VertexBuffers;
-use mapr::benchmarking::io::static_tile_fetcher::StaticTileFetcher;
-use mapr::benchmarking::tessellation::Tessellated;
+use maplibre::benchmarking::io::static_tile_fetcher::StaticTileFetcher;
+use maplibre::benchmarking::tessellation::Tessellated;
 use std::io::Cursor;
 
 const MUNICH_X: u32 = 17425;
