+++ docs/book.toml
@@ -3,4 +3,4 @@ authors = ["Maximilian Ammann"]
 language = "en"
 multilingual = false
 src = "src"
-title = "mapr documentation"
+title = "maplibe-rs documentation"
