+++ docs/src/developer-log.md
@@ -1,3 +1,3 @@
 # Developer Log
 
-I'm regularly releasing blog posts [on my blog](https://maxammann.org/categories/mapr/).
+I'm regularly releasing blog posts [on my blog](https://maxammann.org/categories/maplibre/).
