+++ docs/src/development-documents/caching.md
@@ -1,7 +1,7 @@
 # Caching
 
-The caching for mapr is handled on the networking layer. This means that data which is fetched over slow IO is cached in
-the format of the network requests. The mapr library is not introducing a separate serialization format for caching.
+The caching for maplibre-rs is handled on the networking layer. This means that data which is fetched over slow IO is cached in
+the format of the network requests. The maplibre-rs library is not introducing a separate serialization format for caching.
 
 Instead, caching functionality of HTTP client libraries of the web platform are used. This has the advantage that we can
 honor HTTP headers which configure caching. This is very important for fetched tiles, as they can have an expiry date.
