+++ web/.gitignore
@@ -1,5 +1,5 @@
 node_modules
 
-libs/mapr*
+libs/maplibre*
 
 dist
