+++ web/src/index.ejs
@@ -4,6 +4,6 @@
  <title><%= htmlWebpackPlugin.options.title %></title>
 </head>
 <body style="margin: 0; padding: 0; width:100%; height: 100%; max-height: 100%;">
-<canvas id="mapr" style="position:absolute;"></canvas>
+<canvas id="maplibre" style="position:absolute;"></canvas>
 </body>
 </html>
